Revert "Some changes"
This reverts commit acdaa722d98d25ed79a10d16609b160c2121eb47.

+++ src/BossEnemy.java
@@ -2,7 +2,7 @@ import java.awt.*;
 
 public class BossEnemy extends Enemy {
 
-    public BossEnemy(double health, double maxHealth, double damage, double armor, int width, int height, Image img) {
+    public BossEnemy(float health, float maxHealth, float damage, float armor, int width, int height, Image img) {
         super(health, maxHealth, damage, armor, width, height, 0, img);
         x = GameFrame.WIDTH / 2 - width / 2;
         y = GameFrame.HEIGHT / 2 - height / 2;
